# core/openapi.yaml
openapi: 3.1.0
info:
  title: Madang core API
  version: 2.0.0
  summary: madang-core의 로컬 HTTP·WebSocket API.
  description: |
    데스크톱 앱은 이 API로 madang-core와 통신한다. 페이지 기록은 프로젝트
    폴더의 `.madang/`에, 전역 설정(config.yaml)과 profile.md는 앱 홈(`~/.madang`)에 있다.
    둘 다 core가 쓰며 앱은 자체 문서를 두지 않는다. core는 기록을 git에
    커밋하지 않는다(커밋 여부는 사용자가 정한다).

    규약:
    - 시각은 오프셋이 붙은 ISO 8601이다(`2026-09-24T09:40:00+09:00`).
    - 페이지 id는 `YYYY-MM-DD-<slug>`, 블록 id는 `bNN`이다. run 번호는
      페이지별로 늘어나며 재사용하지 않는다.
    - 페이지 안의 경로는 페이지 기준이다(`blocks/b05-base.json`). 프로젝트
      폴더 기준 경로에는 `repo:` 접두어가 붙는다
      (`repo:src/session/lock.ts`).
    - 오류는 `ApiError`로 돌려준다. 파일 검사에 실패하면 400과 함께
      `ValidationFailure`와 그 `issues`를 돌려준다.
    - 변경은 `/events` WebSocket으로 알린다. 앱은 재연결 후 전부 다시
      가져온다.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://127.0.0.1:7470
    description: 로컬 core. 포트는 다를 수 있으며 core가 core.port에 적는다.
...
paths:
...
  /config/routes:
    get:
      tags: [setup]
      operationId: getRoutes
      summary: config.yaml의 routes 절 원문.
      responses:
        "200":
          description: 라우팅 표.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutesDocument"
        "409":
          $ref: "#/components/responses/Conflict"
    put:
      tags: [setup]
      operationId: saveRoutes
      summary: config.yaml의 routes 절을 바꾼다. 검사를 통과해야 저장된다.
      description: |
        YAML 문법, 필수 키(`kinds`, `default_kind`, `tiers`), 스키마, 종류와
        runner 참조를 검사한다. 실패하면 파일은 그대로다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoutesDocument"
      responses:
        "200":
          description: 저장된 라우팅 표.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutesDocument"
        "400":
          $ref: "#/components/responses/Invalid"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /pages/{page}/messages:
    parameters:
      - $ref: "#/components/parameters/Page"
    post:
      tags: [messages]
      operationId: sendMessage
      summary: 사용자 메시지를 추가하고 run을 시작한다.
      description: |
        run은 `run.*` 이벤트로 알린다. 선택한 요소를 겨냥한 요청
        (`mode: edit`과 `elements`)은 항상 종류 `small`로 실행한다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageCreate"
      responses:
        "202":
          description: 메시지를 저장하고 flow를 시작했다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageAccepted"
        "400":
          $ref: "#/components/responses/Invalid"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /pages/{page}/runs/{n}:
    parameters:
      - $ref: "#/components/parameters/Page"
      - $ref: "#/components/parameters/RunNumber"
    get:
      tags: [runs]
      operationId: getRun
      summary: run 기록(runs/N.json).
      responses:
        "200":
          description: 기록.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunRecord"
        "404":
          $ref: "#/components/responses/NotFound"
  /pages/{page}/runs/{n}/cancel:
    parameters:
      - $ref: "#/components/parameters/Page"
      - $ref: "#/components/parameters/RunNumber"
    post:
      tags: [runs]
      operationId: cancelRun
      summary: 진행 중인 run을 멈춘다. `cancelled`로 끝난다.
      responses:
        "202":
          description: 멈추는 중. 이어서 `run.failed`가 온다.
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /pages/{page}/runs/{n}/events:
    parameters:
      - $ref: "#/components/parameters/Page"
      - $ref: "#/components/parameters/RunNumber"
    get:
      tags: [runs]
      operationId: listRunEvents
      summary: run의 이벤트 로그(runs/N.jsonl).
      description: |
        runner 이벤트를 도착 순서대로 돌려준다. 진행 중인 run이면 그때까지의
        이벤트다. 레이어 1의 run 패널이 읽는다.
      responses:
        "200":
          description: run의 이벤트.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RunStreamEvent"
                examples:
                  - - type: text
                      text: work 배열에서 펄어비스 항목을 찾습니다.
                    - type: tool_call
                      name: apply_patch
                      summary: blocks/b05-base.json
                    - type: file_changed
                      path: blocks/b05-base.json
                    - type: usage
                      usage: {input: 29104, cached: 24310, output: 612}
                    - type: done
                      text: work 순서를 바꿨습니다. 펄어비스 경력이 맨 위입니다.
        "404":
          $ref: "#/components/responses/NotFound"
  /pages/{page}/runs/{n}/undo:
    parameters:
      - $ref: "#/components/parameters/Page"
      - $ref: "#/components/parameters/RunNumber"
    post:
      tags: [runs]
      operationId: undoRun
      summary: run 하나의 부작용을 역순으로 되감는다("되돌리기").
      description: |
        recorder가 남긴 `runs/N.undo.json`대로 게시(게시 기록), git 커밋·머지
        (되돌림 커밋, `reset`을 쓰지 않는다), 페이지 파일(실행 전 사본)을
        나중 것부터 되감는다. 그 run 뒤에 파일이 다시 바뀌었거나, 커밋이
        이력에서 사라졌거나, 더 나중 게시가 있거나, 이미 되돌렸으면 아무것도
        바꾸지 않고 409. 흐름이 돌고 있으면 409 `busy`. 끝나면
        `page.updated`, 되돌림 커밋마다 `git.changed`, 되감은 게시마다
        `publish.done`(`undo: true`)을 낸다.
      responses:
        "200":
          description: 되감은 것.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UndoResult"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /pages/{page}/asks/{id}/answer:
    parameters:
      - $ref: "#/components/parameters/Page"
      - name: id
        in: path
        required: true
        description: page.md 묻는 블록의 id(`ask.created`의 `block`).
        schema:
          type: string
        example: b12
    post:
      tags: [decisions]
      operationId: answerAsk
      summary: 정책이 멈춘 자리의 묻는 블록에 답하고 흐름을 이어 간다.
      description: |
        정책(테스트·충돌·금지 명령·게시 실패)이 부작용을 멈추면 page.md에 묻는
        블록이 생기고 사람은 거기서만 답한다. 답은 page.md에 블록으로 남는다.
        선택지(`merge`, `retry`, `stop` 등)에 없는 답은 400, 그 블록을 기다리는
        흐름이 없으면 404, 흐름이 돌고 있으면 409 `busy`. 같은 결정은
        `POST /pages/{page}/decisions/{id}/answer`로도 답할 수 있다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecisionAnswer"
            example:
              choice: merge
      responses:
        "202":
          description: 받았다. 흐름 이벤트가 이어진다.
        "400":
          $ref: "#/components/responses/Invalid"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /pages/{page}/decisions/{id}/answer:
    parameters:
      - $ref: "#/components/parameters/Page"
      - name: id
        in: path
        required: true
        description: "`flow.waiting`으로 받은 대기 중 결정의 id."
        schema:
          type: string
    post:
      tags: [decisions]
      operationId: answerDecision
      summary: 사람 결정에 답해 기다리던 flow를 재개한다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecisionAnswer"
      responses:
        "202":
          description: flow를 재개했다.
        "400":
          $ref: "#/components/responses/Invalid"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /pages/{page}/preview-input:
    parameters:
      - $ref: "#/components/parameters/Page"
    get:
      tags: [messages]
      operationId: previewInput
      summary: 다음 호출이 읽을 내용과 부분별 토큰 추정치.
      description: 앱은 사용자가 입력하는 동안 디바운스해서 호출한다.
      parameters:
        - name: target
          in: query
          required: false
          description: 대상 블록 id. 생략하면 페이지를 대상으로 한다.
          schema:
            type: string
          example: b04
        - name: elements
          in: query
          required: false
          description: 선택한 요소의 데이터 경로.
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
          example: ["work[1]"]
        - name: mode
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TargetMode"
        - name: text
          in: query
          required: false
          description: 지금까지 입력한 요청 문장.
          schema:
            type: string
      responses:
        "200":
          description: 입력 내역.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InputPreview"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
...
    Page:
      name: page
      in: path
      required: true
      description: 페이지 id.
      schema:
        type: string
      example: 2026-09-24-resume
...
    RunNumber:
      name: n
      in: path
      required: true
      description: 페이지 안의 run 번호.
      schema:
        type: integer
        minimum: 1
      example: 2
...
  responses:
    Invalid:
      description: 요청 또는 결과 파일이 검사에 실패했다.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationFailure"
    NotFound:
      description: 프로젝트, 페이지, 블록, run, 파일이 없다.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error: not_found
            message: page '2026-09-30-missing' not found
    Conflict:
      description: 현재 상태에서는 변경할 수 없다.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error: conflict
            message: "project 'jobs' still has child projects: jobs-2025"
  schemas:
...
    ApiError:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: >-
            기계가 읽는 사유: `not_found`, `conflict`, `invalid`, `busy`,
            `no_repo`, `denied`(정책의 금지 명령), 게시 오류 코드
            (`target-missing`, `nothing-to-undo`, `target-changed` 등).
        message:
          type: string
      examples:
        - error: conflict
          message: a run is already in progress on this page
        - error: denied
          message: "금지 명령: push --force"
    Issue:
      type: object
      description: 검사 문제 하나.
      required: [code, message]
      properties:
        code:
          type: string
          description: >-
            케밥 표기 식별자. 알려진 코드: `missing-file`,
            `unreadable`, `frontmatter`, `missing-key`, `invalid-value`,
            `duplicate-decision`, `choice-not-in-options`, `repo-unset`,
            `invalid-artifact`, `artifact-missing`, `missing-section`,
            `token-limit`, `done-without-verify`, `invalid-yaml`.
        message:
          type: string
        line:
          type: [integer, "null"]
          description: 1부터 세는 파일 줄. 알 때만.
        path:
          type: [string, "null"]
          description: 파일. 알 때만.
      examples:
        - code: missing-section
          message: "missing required section '## 다음 할 일'"
          line: null
          path: ledger.md
    ValidationFailure:
      type: object
      required: [error, message, issues]
      properties:
        error:
          type: string
          enum: [invalid]
        message:
          type: string
        issues:
          type: array
          items:
            $ref: "#/components/schemas/Issue"
      examples:
        - error: invalid
          message: ledger.md failed validation
          issues:
            - code: invalid-value
              message: "status: expected planning | doing | blocked | review | done"
              line: 2
              path: ledger.md
            - code: token-limit
              message: ledger.md is 2310 tokens (limit 2000); summarize the log
              line: null
              path: ledger.md

    # 시스템
...
    RoutesDocument:
      type: object
      description: config.yaml의 routes 절 원문.
      required: [text]
      properties:
        text:
          type: string
      examples:
        - text: |
            kinds: [design, build, small, review, explore]
            default_kind: build
            tiers:
              design: [{runner: claude, model: claude-opus-5-5, effort: high}]
              build: [{runner: codex, model: gpt-6-sol, effort: medium}]
              small: [{runner: codex, model: gpt-6-luna, effort: high}]
              review: [{runner: opposite, model: primary}]
              explore: [{runner: codex, model: gpt-6-luna, effort: medium}]

    # 러너
...
    TargetMode:
      type: string
      enum: [view, edit]
    Target:
      type: object
      description: 메시지가 향하는 대상. 비어 있으면 페이지.
      properties:
        block:
          type: string
        elements:
          type: array
          description: "요소 데이터 경로. 예: `work[1].company`."
          items:
            type: string
        mode:
          $ref: "#/components/schemas/TargetMode"
      examples:
        - block: b04
          elements: ["work[1]"]
          mode: edit
...
    MessageCreate:
      type: object
      required: [text]
      properties:
        text:
          type: string
          description: "`design:` 같은 접두어가 종류를 강제한다."
        target:
          $ref: "#/components/schemas/Target"
      examples:
        - text: 펄어비스 경력을 맨 위로
          target: {block: b04, elements: ["work[1]"], mode: edit}
    MessageAccepted:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: 저장된 메시지 블록의 id.
      examples:
        - message: b07
    InputParts:
      type: object
      description: 입력 부분별 추정 토큰.
      required: [system_est, profile, brief, ledger, contract, target, request]
      properties:
        system_est:
          type: integer
          description: runner 자체의 시스템 프롬프트와 도구(상수).
        profile:
          type: integer
          description: profile.md 본문.
        brief:
          type: integer
          description: brief.md 본문.
        ledger:
          type: integer
          description: ledger.md(머리부와 본문).
        contract:
          type: integer
          description: 내장 작업 지시문.
        target:
          type: integer
        request:
          type: integer
      examples:
        - {system_est: 24600, profile: 110, brief: 1840, ledger: 1320, contract: 420, target: 900, request: 40}
    RunInput:
      type: object
      required: [parts, total_est]
      properties:
        parts:
          $ref: "#/components/schemas/InputParts"
        total_est:
          type: integer
      examples:
        - parts: {system_est: 24600, profile: 110, brief: 1840, ledger: 1320, contract: 420, target: 900, request: 40}
          total_est: 29230
    InputPreview:
      type: object
      description: 다음 호출이 어떻게 라우팅되고 무엇을 읽는지.
      required: [kind, tier, runner, model, parts, total_est]
      properties:
        kind:
          type: string
        tier:
          type: integer
        runner:
          type: string
        model:
          type: string
        effort:
          type: string
        parts:
          $ref: "#/components/schemas/InputParts"
        total_est:
          type: integer
      examples:
        - kind: small
          tier: 1
          runner: codex
          model: gpt-6-luna
          effort: high
          parts: {system_est: 24600, profile: 110, brief: 1840, ledger: 1320, contract: 420, target: 900, request: 40}
          total_est: 29230

    # 실행
    RunResultStatus:
      type: string
      description: >-
        runner가 끝난 방식(`error`, `cancelled`, 시간 초과는 `blocked`), 또는
        정상 종료 뒤 ledger.md의 status.
      enum: [planning, doing, blocked, review, done, error, cancelled]
    RunTrigger:
      type: object
      properties:
        message:
          type: string
        target:
          type: string
        elements:
          type: array
          items:
            type: string
        mode:
          $ref: "#/components/schemas/TargetMode"
      examples:
        - {message: b07, target: b04, elements: ["work[1]"], mode: edit}
    RunUsage:
      type: object
      description: 토큰 수. `input`은 `cached`를 포함한다.
      required: [input, cached, output]
      properties:
        input:
          type: integer
        cached:
          type: integer
        output:
          type: integer
      examples:
        - {input: 29104, cached: 24310, output: 612}
    RunVerify:
      type: object
      properties:
        cmd:
          type: [string, "null"]
        ok:
          type: [boolean, "null"]
      examples:
        - {cmd: null, ok: null}
    RunRecord:
      type: object
      description: runs/N.json 형식의 run 기록.
      required: [n, usage, changed_files, unknown_files, verify]
      properties:
        n:
          type: integer
        started:
          type: [string, "null"]
          format: date-time
        finished:
          type: [string, "null"]
          format: date-time
        trigger:
          $ref: "#/components/schemas/RunTrigger"
        kind:
          type: string
        tier:
          type: integer
        runner:
          type: string
        model:
          type: string
        effort:
          type: string
        input:
          $ref: "#/components/schemas/RunInput"
        usage:
          $ref: "#/components/schemas/RunUsage"
        changed_files:
          type: array
          items:
            type: string
        unknown_files:
          type: array
          items:
            type: string
        verify:
          $ref: "#/components/schemas/RunVerify"
        result_status:
          $ref: "#/components/schemas/RunResultStatus"
        events_log:
          type: [string, "null"]
          description: 원본 이벤트 스트림의 페이지 기준 경로.
      examples:
        - n: 2
          started: "2026-09-24T09:31:00+09:00"
          finished: "2026-09-24T09:31:41+09:00"
          trigger: {message: b07, target: b04, elements: ["work[1]"], mode: edit}
          kind: small
          tier: 1
          runner: codex
          model: gpt-6-luna
          effort: high
          input:
            parts: {system_est: 24600, profile: 110, brief: 1840, ledger: 1320, contract: 420, target: 900, request: 40}
            total_est: 29230
          usage: {input: 29104, cached: 24310, output: 612}
          changed_files: [blocks/b05-base.json]
          unknown_files: []
          verify: {cmd: null, ok: null}
          result_status: review
          events_log: runs/2.jsonl
    RunStreamEvent:
      type: object
      description: >-
        도구에 중립인 runner 이벤트 하나. `type`에 해당하는 필드만 채운다:
        text는 `text`, tool_call은 `name`·`summary`, tool_result는
        `summary`, file_changed는 `path`, usage는 `usage`, done은 `text`,
        error는 `message`.
      required: [type]
      properties:
        type:
          type: string
          enum: [text, tool_call, tool_result, file_changed, usage, done, error]
        text:
          type: string
        name:
          type: string
        summary:
          type: string
        path:
          type: string
        usage:
          $ref: "#/components/schemas/RunUsage"
        message:
          type: string
      examples:
        - type: tool_call
          name: apply_patch
          summary: blocks/b05-base.json
        - type: file_changed
          path: blocks/b05-base.json

    # 결정
...
    DecisionAnswer:
      type: object
      required: [choice]
      properties:
        choice:
          type: string
          description: >-
            `choice`는 선택지 중 하나, `yesno`는 `true`/`false`, `score`는
            `1`–`5`.
      examples:
        - choice: retry

    # 상태(에이전트)
...
    UndoResult:
      type: object
      description: run 하나를 되감은 결과.
      required: [run, restored, skipped, reverted, unpublished]
      properties:
        run:
          type: integer
        restored:
          type: array
          description: 실행 전 내용으로 되감은 페이지 파일(페이지 기준).
          items:
            type: string
        skipped:
          type: array
          description: 되감지 않는 부작용(`repo:` 작업 트리 파일 등). 커밋·머지 되돌림이 대신 되감는다.
          items:
            type: string
        reverted:
          type: array
          description: git 커밋·머지를 되감은 되돌림 커밋 해시.
          items:
            type: string
        unpublished:
          type: array
          description: 되감은 게시 번호.
          items:
            type: integer
      examples:
        - run: 3
          restored: [ledger.md, page.md]
          skipped: ["repo:src/session/lock.ts"]
          reverted: [9f8e7d6c5b4a39281706f5e4d3c2b1a098f7e6d5]
          unpublished: []

    # 게시
...